backend/artifact-identifier-service/requirements.txt
flask>=3.1.2
flask-cors>=6.0.1
python-dotenv>=1.1.1
tensorflow>=2.18.0
numpy>=1.24.3
pandas>=2.3.3
openpyxl>=3.1.5
pillow>=10.0.0
werkzeug>=3.1.3
requests>=2.31.0
gdown>=5.1.0
joblib>=1.3.0
scikit-learn>=1.3.0
